Move CreateOrUpdateConfiguration
